feat: dynamic tokens per network achieved
