+++ autotx/cli.py
@@ -1,7 +1,7 @@
 import click
 from dotenv import load_dotenv
 
-from autotx.utils.ethereum.constants import SUPPORTED_NETWORKS
+from autotx.utils.ethereum.networks import NetworkInfo
 from autotx.utils.ethereum.eth_address import ETHAddress
 from autotx.utils.ethereum.helpers.get_dev_account import get_dev_account
 load_dotenv()
@@ -32,10 +32,7 @@ def run(prompt: str, non_interactive: bool):
     chain_id = web3.eth.chain_id
     print(f"Chain ID: {chain_id}")
 
-    network_info = SUPPORTED_NETWORKS.get(chain_id)
-
-    if network_info is None:
-        raise Exception(f"Chain ID {chain_id} is not supported")
+    network_info = NetworkInfo(chain_id)
 
     manager: SafeManager
 
@@ -57,9 +54,6 @@ def run(prompt: str, non_interactive: bool):
         send_eth(dev_account, manager.address, 10, web3)
         print(f"Sent 10 ETH to smart account for testing purposes")
 
-    print("Starting smart account balances:")
-    show_address_balances(web3, network_info, manager.address)
-
     autotx = AutoTx(manager, network_info, [
         SendTokensAgent.build_agent_factory(),
         SwapTokensAgent.build_agent_factory(client, manager.address),
@@ -67,8 +61,6 @@ def run(prompt: str, non_interactive: bool):
     ], None)
     autotx.run(prompt, non_interactive)
 
-    print("Final smart account balances:")
-    show_address_balances(web3, network_info, manager.address)
 
 @main.group()
 def agent():
