+++ autotx/tests/token/test_swap.py
@@ -1,6 +1,6 @@
 from autotx.patch import patch_langchain
 from autotx.utils.ethereum import load_w3
-from autotx.utils.ethereum.constants import SUPPORTED_NETWORKS
+from autotx.utils.ethereum.networks import SUPPORTED_NETWORKS
 from autotx.utils.ethereum.eth_address import ETHAddress
 
 patch_langchain()
